clean up and move back to module structure
